polar: fixed channel construction naming error
